requirements.txt
numpy==1.21.5
opencv_python_headless==4.7.0.72
pandas==1.4.4
Pillow==9.2.0
Pillow==10.0.1
scikit_learn==1.0.2
timm==0.6.12
torch==1.12.1
torchvision==0.13.1
